move visuals into a separate folder
