+++ src/utils/util_visuals.js
@@ -2,10 +2,10 @@
 
 // import { getCPUCost } from "@/utils/util_helper"
 // import roleTagger from "./visuals_roleTagger"
-import roomBuildingPlanner from "./visuals_roomBuildingPlanner"
-import { showRoomRoleCounts } from "./visuals_roomRoleCounts"
-import { showRoomSpawnQueue } from "./visuals_roomSpawnQueue"
-import worldVisual_showMineralMap from "./visuals_world_showMineralMap"
+import roomBuildingPlanner from "./visual/visuals_roomBuildingPlanner"
+import { showRoomRoleCounts } from "./visual/visuals_roomRoleCounts"
+import { showRoomSpawnQueue } from "./visual/visuals_roomSpawnQueue"
+import worldVisual_showMineralMap from "./visual/visuals_world_showMineralMap"
 
 const showVisuals = () => {
