+++ src/utils/visual/visuals_roomBuildingPlanner.js
