+++ src/utils/visual/visuals_world_showMineralMap.js
